feat: Implement authentication and authorization for events and attendees management

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -6,16 +6,28 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\AttendeeResource;
 use App\Models\Event;
 use App\Models\Attendee;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
+use App\Http\Traits\CanLoadRelationships;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 
-class AttendeeController extends Controller
+class AttendeeController extends Controller implements HasMiddleware
 {
-    /**
-     * Display a listing of the resource.
-     */
+    use CanLoadRelationships, AuthorizesRequests;
+
+    private array $relations = ['user'];
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show']),
+        ];
+    }
+
     public function index(Event $event)
     {
-        $attendees = $event->attendees()->latest();
+        $attendees = $this->loadRelationships($event->attendees()->latest());
 
 
         return AttendeeResource::collection(
@@ -28,10 +40,12 @@ class AttendeeController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request , Event $event)
-   {
+    public function store(Request $request, Event $event)
+    {
+        $this->authorize('create', Attendee::class);
+
         $attendee = $event->attendees()->create([
-            'user_id' => 1
+            'user_id' => $request->user()->id
         ]);
 
         return new AttendeeResource($attendee);
@@ -42,7 +56,7 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee)
     {
-        return new AttendeeResource($attendee);
+        return new AttendeeResource($this->loadRelationships($attendee));
     }
 
     /**
@@ -58,6 +72,9 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete', $attendee);
+
+         $attendee = $event->attendees()->where('id', $attendee->id)->firstOrFail();
         $attendee->delete();
         return response(status: 204);
     }
